First Unit Test in repository; fixed errors; Structural Change to CMakeLists
Adds the first unit test to check the distance calculations for different skeletons (in Distancematrixoperations.hpp). In the process found errors in DistanceMatrixoperations that were fixed.
Also changes the structure of CMakelists to add most files of an executable in a libraray except for the ros interactor. This way there is a clear separation between the binaries with the ros wrapper and the other files that constitute the 'core logic'. Necessary to make unit testing flexible with CMakeLists (only need to add / remove files in one place in the CMakeLists file: in the library).