Workaround for the bug in the Munkres Library
The check validity method of the library does not work properly and throws exceptions when run. It is now simply not used, the check that was run should never fail anyway. Note that there is still the check for validity of matrix format that is run (written in this framework, not in the Munkres library).